+++ git/git.go
@@ -19,6 +19,9 @@ type Gitter interface {
 	SetOriginWithCredentials(scm.Repo) error
 	Clean(scm.Repo) error
 	Checkout(scm.Repo) error
+	RevListCompare(scm.Repo, string, string) (string, error)
+	ShortStatus(scm.Repo) (string, error)
+	Branch(scm.Repo) (string, error)
 	UpdateRemote(scm.Repo) error
 	FetchAll(scm.Repo) error
 	FetchCloneBranch(scm.Repo) error
@@ -182,6 +185,35 @@ func (g GitClient) FetchAll(repo scm.Repo) error {
 	return cmd.Run()
 }
 
+func (g GitClient) Branch(repo scm.Repo) (string, error) {
+	args := []string{"branch"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = repo.HostPath
+	if os.Getenv("GHORG_DEBUG") != "" {
+		if err := printDebugCmd(cmd, repo); err != nil {
+			return "", err
+		}
+	}
+
+	output, err := cmd.Output()
+	if err != nil {
+		return "", err
+	}
+
+	return strings.TrimSpace(string(output)), nil
+}
+
+// RevListCompare returns the list of commits in the local branch that are not in the remote branch.
+func (g GitClient) RevListCompare(repo scm.Repo, localBranch string, remoteBranch string) (string, error) {
+	cmd := exec.Command("git", "-C", repo.HostPath, "rev-list", localBranch, "^"+remoteBranch)
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(string(output)), nil
+}
+
 func (g GitClient) FetchCloneBranch(repo scm.Repo) error {
 	args := []string{"fetch", "origin", repo.CloneBranch}
 
@@ -196,6 +228,25 @@ func (g GitClient) FetchCloneBranch(repo scm.Repo) error {
 		return printDebugCmd(cmd, repo)
 	}
 	return cmd.Run()
+}
+
+func (g GitClient) ShortStatus(repo scm.Repo) (string, error) {
+	args := []string{"status", "--short"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = repo.HostPath
+	if os.Getenv("GHORG_DEBUG") != "" {
+		if err := printDebugCmd(cmd, repo); err != nil {
+			return "", err
+		}
+	}
+
+	output, err := cmd.Output()
+	if err != nil {
+		return "", err
+	}
+
+	return strings.TrimSpace(string(output)), nil
 }
 
 func (g GitClient) RepoCommitCount(repo scm.Repo) (int, error) {